backend/requirements.txt
Django==5.0
psycopg[binary,pool]
python-dotenv==1.0.0
djangorestframework==3.14.0
django-allauth==0.59.0
dj-rest-auth==5.0.2
